+++ src/compile/nodes/AwaitBlock.ts
@@ -1,10 +1,7 @@
-import deindent from '../../utils/deindent';
 import Node from './shared/Node';
-import Block from '../render-dom/Block';
 import PendingBlock from './PendingBlock';
 import ThenBlock from './ThenBlock';
 import CatchBlock from './CatchBlock';
-import createDebuggingComment from '../../utils/createDebuggingComment';
 import Expression from './shared/Expression';
 
 export default class AwaitBlock extends Node {
@@ -28,169 +25,5 @@ export default class AwaitBlock extends Node {
 		this.pending = new PendingBlock(component, this, scope, info.pending);
 		this.then = new ThenBlock(component, this, scope.add(this.value, deps), info.then);
 		this.catch = new CatchBlock(component, this, scope.add(this.error, deps), info.catch);
-	}
-
-	init(
-		block: Block,
-		stripWhitespace: boolean,
-		nextSibling: Node
-	) {
-		this.cannotUseInnerHTML();
-
-		this.var = block.getUniqueName('await_block');
-		block.addDependencies(this.expression.dependencies);
-
-		let isDynamic = false;
-		let hasIntros = false;
-		let hasOutros = false;
-
-		['pending', 'then', 'catch'].forEach(status => {
-			const child = this[status];
-
-			child.block = block.child({
-				comment: createDebuggingComment(child, this.component),
-				name: this.component.getUniqueName(`create_${status}_block`)
-			});
-
-			child.initChildren(child.block, stripWhitespace, nextSibling);
-			this.component.target.blocks.push(child.block);
-
-			if (child.block.dependencies.size > 0) {
-				isDynamic = true;
-				block.addDependencies(child.block.dependencies);
-			}
-
-			if (child.block.hasIntros) hasIntros = true;
-			if (child.block.hasOutros) hasOutros = true;
-		});
-
-		this.pending.block.hasUpdateMethod = isDynamic;
-		this.then.block.hasUpdateMethod = isDynamic;
-		this.catch.block.hasUpdateMethod = isDynamic;
-
-		this.pending.block.hasIntroMethod = hasIntros;
-		this.then.block.hasIntroMethod = hasIntros;
-		this.catch.block.hasIntroMethod = hasIntros;
-
-		this.pending.block.hasOutroMethod = hasOutros;
-		this.then.block.hasOutroMethod = hasOutros;
-		this.catch.block.hasOutroMethod = hasOutros;
-
-		if (hasOutros && this.component.options.nestedTransitions) block.addOutro();
-	}
-
-	build(
-		block: Block,
-		parentNode: string,
-		parentNodes: string
-	) {
-		const name = this.var;
-
-		const anchor = this.getOrCreateAnchor(block, parentNode, parentNodes);
-		const updateMountNode = this.getUpdateMountNode(anchor);
-
-		const { snippet } = this.expression;
-
-		const info = block.getUniqueName(`info`);
-		const promise = block.getUniqueName(`promise`);
-
-		block.addVariable(promise);
-
-		block.maintainContext = true;
-
-		const infoProps = [
-			block.alias('component') === 'component' ? 'component' : `component: #component`,
-			'ctx',
-			'current: null',
-			this.pending.block.name && `pending: ${this.pending.block.name}`,
-			this.then.block.name && `then: ${this.then.block.name}`,
-			this.catch.block.name && `catch: ${this.catch.block.name}`,
-			this.then.block.name && `value: '${this.value}'`,
-			this.catch.block.name && `error: '${this.error}'`,
-			this.pending.block.hasOutroMethod && `blocks: Array(3)`
-		].filter(Boolean);
-
-		block.builders.init.addBlock(deindent`
-			let ${info} = {
-				${infoProps.join(',\n')}
-			};
-		`);
-
-		block.builders.init.addBlock(deindent`
-			@handlePromise(${promise} = ${snippet}, ${info});
-		`);
-
-		block.builders.create.addBlock(deindent`
-			${info}.block.c();
-		`);
-
-		if (parentNodes) {
-			block.builders.claim.addBlock(deindent`
-				${info}.block.l(${parentNodes});
-			`);
-		}
-
-		const initialMountNode = parentNode || '#target';
-		const anchorNode = parentNode ? 'null' : 'anchor';
-
-		const hasTransitions = this.pending.block.hasIntroMethod || this.pending.block.hasOutroMethod;
-
-		block.builders.mount.addBlock(deindent`
-			${info}.block.${hasTransitions ? 'i' : 'm'}(${initialMountNode}, ${info}.anchor = ${anchorNode});
-			${info}.mount = () => ${updateMountNode};
-		`);
-
-		const conditions = [];
-		if (this.expression.dependencies.size > 0) {
-			conditions.push(
-				`(${[...this.expression.dependencies].map(dep => `'${dep}' in changed`).join(' || ')})`
-			);
-		}
-
-		conditions.push(
-			`${promise} !== (${promise} = ${snippet})`,
-			`@handlePromise(${promise}, ${info})`
-		);
-
-		block.builders.update.addLine(
-			`${info}.ctx = ctx;`
-		);
-
-		if (this.pending.block.hasUpdateMethod) {
-			block.builders.update.addBlock(deindent`
-				if (${conditions.join(' && ')}) {
-					// nothing
-				} else {
-					${info}.block.p(changed, @assign(@assign({}, ctx), ${info}.resolved));
-				}
-			`);
-		} else {
-			block.builders.update.addBlock(deindent`
-				${conditions.join(' && ')}
-			`);
-		}
-
-		if (this.pending.block.hasOutroMethod && this.component.options.nestedTransitions) {
-			const countdown = block.getUniqueName('countdown');
-			block.builders.outro.addBlock(deindent`
-				const ${countdown} = @callAfter(#outrocallback, 3);
-				for (let #i = 0; #i < 3; #i += 1) {
-					const block = ${info}.blocks[#i];
-					if (block) block.o(${countdown});
-					else ${countdown}();
-				}
-			`);
-		}
-
-		block.builders.destroy.addBlock(deindent`
-			${info}.block.d(${parentNode ? '' : 'detach'});
-			${info} = null;
-		`);
-
-		[this.pending, this.then, this.catch].forEach(status => {
-			status.children.forEach(child => {
-				child.build(status.block, null, 'nodes');
-			});
-		});
 	}
 }
