+++ src/compile/render-dom/wrappers/EachBlock.ts
@@ -1,7 +1,7 @@
 import Renderer from '../Renderer';
 import Block from '../Block';
 import Node from '../../nodes/shared/Node';
-import Wrapper from './shared/wrapper';
+import Wrapper from './shared/Wrapper';
 import createDebuggingComment from '../../../utils/createDebuggingComment';
 import EachBlock from '../../nodes/EachBlock';
 import FragmentWrapper from './Fragment';
